Bug 820188 Email app integration tests a=test-only r=squib
Implements two integration tests for the Email app. One that covers
basic launch and first account setup using an example.com address
(which is a fake account) and the other allows you to run
through the launch using specified credentials.

This includes the integration test scaffolding, support for testvars
for all apps, some helper methods in the EmailIntegration class, and
a README.