Bug 941993 - Change a debug-only pluralized number of seconds string.
It was raised in bug 941993 that we would display "1 seconds" if a value of 1
was used.  This wouldn't happen, but I left the bug open because it's arguably
silly to have the incorrect value.

I am not updating the l10n string since this is a debug only pluralized string.
No user will ever see this string unless they use the secret debug menu to set
their periodic sync value to a ridiculously short value (20 seconds or 60
seconds.)  And in that case the existing localized strings are already correct.

I'm just doing this to close out the bug.